# Connection sorting using Boost's sorting function