--- a/cmake/ProcessOptions.cmake
+++ b/cmake/ProcessOptions.cmake
@@ -356,7 +356,7 @@ function( NEST_PROCESS_WITH_PYTHON )
     # Localize the Python interpreter
     if ( ${with-python} STREQUAL "ON" )
       find_package( PythonInterp )
-    elseif ( ${with-python} STREQUAL "2" )  
+    elseif ( ${with-python} STREQUAL "2" )
       find_package( PythonInterp 2 REQUIRED )
     elseif ( ${with-python} STREQUAL "3" )
       find_package( PythonInterp 3 REQUIRED )
@@ -368,7 +368,7 @@ function( NEST_PROCESS_WITH_PYTHON )
       set( PYTHON ${PYTHON_EXECUTABLE} PARENT_SCOPE )
       set( PYTHON_VERSION ${PYTHON_VERSION_STRING} PARENT_SCOPE )
 
-      # Localize Python lib/header files and make sure that their version matches 
+      # Localize Python lib/header files and make sure that their version matches
       # the Python interpreter version !
       find_package( PythonLibs ${PYTHON_VERSION_STRING} EXACT )
       if ( PYTHONLIBS_FOUND )
@@ -425,7 +425,7 @@ function( NEST_PROCESS_WITH_OPENMP )
       set( CMAKE_C_FLAGS "${CMAKE_C_FLAGS} ${OpenMP_C_FLAGS}" PARENT_SCOPE )
       set( CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${OpenMP_CXX_FLAGS}" PARENT_SCOPE )
     else()
-      message( FATAL_ERROR "CMake can not find OpenMP." ) 
+      message( FATAL_ERROR "CMake can not find OpenMP." )
     endif ()
   endif ()
 endfunction()
@@ -528,7 +528,8 @@ function( NEST_PROCESS_WITH_BOOST )
       set( BOOST_ROOT "${with-boost}" )
     endif ()
 
-    find_package( Boost COMPONENTS unit_test_framework )
+    # Needs Boost version >=1.58.0 to use Boost sorting
+    find_package( Boost 1.58.0 COMPONENTS unit_test_framework )
     if ( Boost_FOUND )
       # export found variables to parent scope
       set( HAVE_BOOST ON PARENT_SCOPE )