--- a/libnestutil/config.h.in
+++ b/libnestutil/config.h.in
@@ -22,10 +22,10 @@
 
 #ifndef CONFIG_H
 #define CONFIG_H
-/* 
+/*
     Configuration header.
     config.h is automatically generated from config.h.in
-    during the configuration process. 
+    during the configuration process.
 
     Please make all changes in config.h.in.
 */
@@ -111,6 +111,9 @@
 
 /* MPI is available. */
 #cmakedefine HAVE_MPI 1
+
+/* Boost is available. */
+#cmakedefine HAVE_BOOST 1
 
 /* MUSIC support enabled? */
 #cmakedefine HAVE_MUSIC 1