--- a/libnestutil/sort.h
+++ b/libnestutil/sort.h
@@ -26,7 +26,15 @@
 #include <cstddef>
 #include <vector>
 
+// Generated includes:
+#include "config.h"
+
 #include "block_vector.h"
+
+#ifdef HAVE_BOOST
+#include <boost/sort/spreadsort/spreadsort.hpp>
+#include "iterator_pair.h"
+#endif
 
 #define INSERTION_SORT_CUTOFF 10 // use insertion sort for smaller arrays
 
@@ -173,7 +181,14 @@ template < typename T1, typename T2 >
 void
 sort( BlockVector< T1 >& vec_sort, BlockVector< T2 >& vec_perm )
 {
+#ifdef HAVE_BOOST
+  boost::sort::spreadsort::integer_sort(
+    make_iterator_pair( vec_sort.begin(), vec_perm.begin() ),
+    make_iterator_pair( vec_sort.end(), vec_perm.end() ),
+    rightshift_iterator_pair() );
+#else
   quicksort3way( vec_sort, vec_perm, 0, vec_sort.size() - 1 );
+#endif
 }
 
 } // namespace sort