This PR implements a helper iterator, `iterator_pair`, to be able to sort contents of two containers (based on values from the first, with permutations applied to the second) using library functions, specifically using the Boost.Sort library. If NEST is not compiled with Boost support, it falls back on the current sorting implementation. The motivation for using a library sorting function is unexpectedly higher sorting times in certain cases, as discussed in #1099.

Note that Boost version >=1.58.0 is required, so if an older version is installed, NEST is installed without Boost.

#### Benchmarks
Addressing the issues discussed in #1099, connecting many neurons to few neurons with `all_to_all` (sorting times given in seconds):

| Connection     | master | Boost.Sort | Speedup |
|----------------|--------|------------|---------|
| 100000->2      | 3.16   | 0.06       | 52.64   |
| 10000->1000    | 5.80   | 0.46       | 12.72   |

Using hpc_benchmark, timing only the sorting:

| Scale | master | Boost.Sort | Speedup |
|-------|--------|------------|---------|
| 8     | 11.26  | 5.42       | 2.08    |
| 12    | 22.82  | 10.61      | 2.15    |
| 16    | 32.92  | 14.19      | 2.32    |
| 20    | 43.88  | 18.58      | 2.36    |

So using the Boost.Sort library both solves the problems discussed in #1099, and sorts (mostly) random connections considerably faster than the current implementation.