@jakobj 2019-02-14
the timing looks very convincing, thanks! haven't looked at it in details, but is doesn't seem to include the fix suggested (https://github.com/nest/nest-simulator/pull/1099#issuecomment-459834317) for the sorting if boost is absent. could you please include this? would be great to see how the timings compare after applying this fix.