@jakobj 2019-02-18
after talking to @suku248 I can express another concern of mine more clearly: the boost implementation does not reduce complexity by replacing the custom sort implementation (necessary as a fallback), but rather increases complexity by introducing a significant amount of new loc via the new `IteratorPair` class. The custom sort and the boost library calls and helpers will hence both need to be maintained. If the performance difference is not an order of magnitude or so, I would hence refrain from increasing complexity and rather have only one implementation.