@heplesser 2019-03-22
@hakonsbm @jakobj Since using the boost sort gives us a factor 2 in realistic cases such as hpc_benchmark, I suggest we use the boost sort if available. At the same time, we should implement alternative pivoting in `nest_quicksort` to avoid the worst-case behavior shown by the two first benchmark results above.