@hakonsbm 2019-03-22
@heplesser I have added the changes proposed by @jakobj.