@gtrensch 2019-03-25
@hakonsbm thanks for this contribution! I also agree with @heplesser , a factor 2 is worth to include boost here. I just deleted an empty line (see the suggested code change above). I approve this PR.